Programming-Standards/python/requirements.txt
Jinja2
PyYAML
pylint>=2.6.0
flake8
black
